requirements.txt
streamlit>=1.36
pandas>=2.2
requests>=2.32
python-dateutil>=2.9
